requirements.txt
numpy~=1.26.2
scipy
pandas~=2.1.3
fiddler-client
Jinja2
scikit-learn
python-docx
docxtpl~=0.16.7
matplotlib~=3.8.2
docx2pdf~=0.1.8

tqdm~=4.66.1
requests~=2.31.0
setuptools~=65.5.1
reportgen~=0.1